requirements.txt
absl-py==0.13.0
apscheduler==3.6.3
astunparse==1.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
chardet==3.0.4
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cs50==7.0.0
cycler==0.10.0
dateparser==1.0.0
distro==1.5.0
flask==2.0.1
flatbuffers==1.12
gast==0.4.0
geographiclib==1.52
geopy==2.2.0
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googlesearch-python==1.0.1
googletrans==3.1.0a0
greenlet==1.1.0
grpcio==1.34.1
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.0.1
jaraco-context==4.0.0
jinja2==3.0.1
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
keras-preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
markdown==3.3.4
markupsafe==2.0.1
matplotlib==3.4.2
more-itertools==8.8.0
neuralintents==0.0.4
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.1
pillow==8.3.1
playsound==1.2.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyaudio==0.2.11
pygame==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.17.0
python-telegram-bot==13.7
pytz==2021.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
speechrecognition==3.8.1
sqlalchemy==1.4.22
sqlparse==0.4.1
squarify==0.4.3
tabula==1.0.5
tabula-py==2.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tornado==6.1
tqdm==4.61.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.6
werkzeug==2.0.1
wolframalpha==5.0.0
wrapt==1.12.1
xmltodict==0.12.0
zalotts==0.0.3
selenium==3.141.0